backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
pandas==2.2.0
openpyxl==3.1.2
google-genai
python-dotenv>=1.0.0
pdfplumber>=0.10.0
